requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
groq
langchain